+++ P2197.cpp
@@ -1,14 +1,14 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-long long main(){
+int main() {
     long long T;
     cin >> T;
-    while(T--){
+    while (T--) {
         long long n;
         cin >> n;
         long long xors = 0, tmp;
-        for (long long i = 0; i < n;i++){
+        for (long long i = 0; i < n;i++) {
             cin >> tmp;
             xors ^= tmp;
         }
