+++ P3155.cpp
@@ -4,9 +4,9 @@ using namespace std;
 long long dp[10005][2];
 vector<long long> maps[10005];
 
-void dfs(long long this_node,long long father_node){
-    for(long long its:maps[this_node]){
-        if(its==father_node){
+void dfs(long long this_node, long long father_node) {
+    for (long long its : maps[this_node]) {
+        if (its == father_node) {
             continue;
         }
         dfs(its, this_node);
@@ -15,21 +15,21 @@ void dfs(long long this_node,long long father_node){
     }
 }
 
-long long main(){
+int main() {
     long long m, n;
     cin >> m >> n;
-    for (long long i = 1; i <= n;i++){
+    for (long long i = 1; i <= n;i++) {
         long long c;
         cin >> c;
         dp[i][c] = 1, dp[i][!c] = long long_MAX;
     }
-    for (long long i = 0; i < m - 1;i++){
+    for (long long i = 0; i < m - 1;i++) {
         long long u, v;
         cin >> u >> v;
         maps[u].push_back(v);
         maps[v].push_back(u);
     }
-    for (long long i = n + 1; i <= m;i++){
+    for (long long i = n + 1; i <= m;i++) {
         dp[i][0] = dp[i][1] = 1;
     }
     dfs(m, -1);
