+++ at_abc394_A.cpp
@@ -1,10 +1,10 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-long long main(){
+int main() {
     char tmp;
-    while(cin>>tmp){
-        if(tmp=='2'){
+    while (cin >> tmp) {
+        if (tmp == '2') {
             cout << 2;
         }
     }
