+++ at_abc394_D.cpp
@@ -3,32 +3,32 @@ using namespace std;
 
 stack<char> ks;
 
-long long main(){
+int main() {
     string s;
     cin >> s;
-    for (long long i = 0; i < s.size();i++){
-        if(s[i]=='(' || s[i]=='[' || s[i]=='<'){
+    for (long long i = 0; i < s.size();i++) {
+        if (s[i] == '(' || s[i] == '[' || s[i] == '<') {
             ks.push(s[i]);
-        }else{
-            if(ks.empty()){
+        } else {
+            if (ks.empty()) {
                 cout << "No";
                 return 0;
             }
-            if(s[i]==')' && ks.top()=='('){
+            if (s[i] == ')' && ks.top() == '(') {
                 ks.pop();
-            }else if(s[i]==']' && ks.top()=='['){
+            } else if (s[i] == ']' && ks.top() == '[') {
                 ks.pop();
-            }else if(s[i]=='>' && ks.top()=='<'){
+            } else if (s[i] == '>' && ks.top() == '<') {
                 ks.pop();
-            }else{
+            } else {
                 cout << "No";
                 return 0;
             }
         }
     }
-    if(ks.empty()){
+    if (ks.empty()) {
         cout << "Yes";
-    }else{
+    } else {
         cout << "No";
     }
     return 0;
