+++ at_abc396_D.cpp
@@ -9,14 +9,14 @@
 
 - 在 $A \oplus B$ 的二进制表示中，当且仅当 $A$ 和 $B$ 的相同数位中恰好有一位是 $1$ 时， $2^k \,(k \ge 0)$ 对应数位上的数字是 $1$ ；否则，它是 $0$ 。
 
-例如， $3 \oplus 5 = 6$ （二进制： $011 \oplus 101 = 110$ ）。  
+例如， $3 \oplus 5 = 6$ （二进制： $011 \oplus 101 = 110$ ）。
 一般来说， $k$ 个整数 $p_1, \dots, p_k$ 的 XOR 定义为 $(\cdots ((p_1 \oplus p_2) \oplus p_3) \oplus \cdots \oplus p_k)$ 。可以证明它与 $p_1, \dots, p_k$ 的阶数无关。
 */
 
 #include <bits/stdc++.h>
 using namespace std;
 
-struct edges{
+struct edges {
     long long v, w;
 };
 
@@ -28,15 +28,15 @@ long long anss;
 
 long long n, m;
 
-void dfs(long long v){
+void dfs(long long v) {
     is_in ^= (1 << v);
-    if(v==n){
+    if (v == n) {
         anss = min(anss, this_v);
         is_in ^= (1 << v);
         return;
     }
-    for (auto i : a[v]){
-        if(!(is_in & (1 << i.v))){
+    for (auto i : a[v]) {
+        if (!(is_in & (1 << i.v))) {
             this_v ^= i.w;
             dfs(i.v);
             this_v ^= i.w;
@@ -45,13 +45,13 @@ void dfs(long long v){
     is_in ^= (1 << v);
 }
 
-long long main(){
+int main() {
     cin >> n >> m;
-    for (long long i = 0; i < m;i++){
+    for (long long i = 0; i < m;i++) {
         long long u, v, w;
         cin >> u >> v >> w;
-        a[u].push_back({v, w});
-        a[v].push_back({u, w});
+        a[u].push_back({ v, w });
+        a[v].push_back({ u, w });
     }
     anss = LLONG_MAX;
     dfs(1);
