+++ at_abc398_C.cpp
@@ -4,17 +4,17 @@ using namespace std;
 unordered_map<long long, long long> s;
 long long sn[300005];
 
-long long main(){
+int main() {
     long long n;
     cin >> n;
-    for (long long i = 0; i < n;i++){
+    for (long long i = 0; i < n;i++) {
         cin >> sn[i];
         s[sn[i]]++;
     }
     long long maxs = -1, anss = -1;
-    for (long long i = n - 1; i >= 0;i--){
-        if(s[sn[i]]==1){
-            if(sn[i]>maxs){
+    for (long long i = n - 1; i >= 0;i--) {
+        if (s[sn[i]] == 1) {
+            if (sn[i] > maxs) {
                 maxs = sn[i];
                 anss = i + 1;
             }
